+++ src/media/py_utils.h
@@ -14,10 +14,10 @@ public:
 	static std::recursive_mutex pyExecMutex;
 
 
-	static PyObject *getCompileObject(const std::string &code, const std::string &fileName, size_t numLine);
+	static PyObject *getCompileObject(const std::string &code, const std::string &fileName, uint32_t numLine);
 	static bool isConstExpr(const std::string &code, bool checkSimple = true);
-	static std::string exec(const std::string &fileName, size_t numLine, const std::string &code, bool retRes = false);
-	static PyObject* execRetObj(const std::string &fileName, size_t numLine, const std::string &code);
+	static std::string exec(const std::string &fileName, uint32_t numLine, const std::string &code, bool retRes = false);
+	static PyObject* execRetObj(const std::string &fileName, uint32_t numLine, const std::string &code);
 	static void errorProcessing(const std::string &code);
 
 	static void init();
